refactor(f): up sanity queries for metadata
